requirements.txt
bibtexparser==1.2.0
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
future==0.18.2
gitdb==4.0.7
GitPython==3.1.17
google-api-core==1.28.0
google-api-python-client==2.5.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
oauth2client==4.1.3
packaging==20.9
protobuf==3.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
smmap==4.0.0
uritemplate==3.0.1
urllib3==1.26.4
click~=8.0.1